Sync pipeline changes from 1.5.4

+++ hack/assets.sh
@@ -3,8 +3,7 @@
 set -eo pipefail
 
 ROOTDIR=$(realpath "${ROOTDIR:-$(dirname "${BASH_SOURCE[0]}")/..}")
-source "${ROOTDIR}/assets.sh"
-source "${ROOTDIR}/common.sh"
+source "${ROOTDIR}/hack/resolve-globs.sh"
 
 if [ $# -ne 1 ] || ([ "${1}" != "--validate" ] && [ "${1}" != "--download" ]); then
     echo "Usage: $0 [--validate|--download]"
@@ -31,7 +30,7 @@ function download_url() {
     echo "ok"
     if [ -n "${auth}" ]; then
         sha256=$(curl -sfSLR -u "${ARTIFACTORY_USER}:${ARTIFACTORY_TOKEN}" "${url/\/artifactory\//\/artifactory\/api\/storage\/}" | jq -r '.checksums.sha256')
-        echo "${sha256}" > "${BUILDDIR}/${path}.sha265.txt"
+        echo "${sha256}" > "${BUILDDIR}/${path}.sha256.txt"
         if ! echo "${sha256} ${BUILDDIR}/${path}" | sha256sum -c --quiet -; then
             echo "SHA256 checksum for downloaded ${path} is incorrect, looks like file was corrupted in transit."
             exit 1
@@ -54,12 +53,25 @@ function process_file() {
         else
             validate_url "${url}" "${auth}"
         fi
+        # images:
+        #   pre-install-toolkit:
+        #     - https://artifactory.algol60.net/artifactory/csm-images/stable/pre-install-toolkit/6.2.30/pre-install-toolkit-6.2.30-x86_64.iso
+        #   kubernetes:
+        # ...
+        dir=$(dirname "${path}")
+        write_version_digest ".${dir//\//.}" "${url}" "${ROOTDIR}/dist/csm-${RELEASE_VERSION}-assets-versions.yaml"
     else
         if [ -n "${CSM_BASE_VERSION}" ]; then
-            if [ -f "${ROOTDIR}/dist/csm-${CSM_BASE_VERSION}/${path}" ]; then
+            if [ -f "${ROOTDIR}/dist/csm-${CSM_BASE_VERSION}/${path}" ] && [ -f "${ROOTDIR}/dist/csm-${CSM_BASE_VERSION}/${path}.sha256.txt" ]; then
                 echo -ne "Found ${path} in CSM base, copying ... "
                 mkdir -p "$(dirname "${BUILDDIR}/${path}")"
                 cp -f "${ROOTDIR}/dist/csm-${CSM_BASE_VERSION}/${path}" "${BUILDDIR}/${path}"
+                cp -f "${ROOTDIR}/dist/csm-${CSM_BASE_VERSION}/${path}.sha256.txt" "${BUILDDIR}/${path}.sha256.txt"
+                sha256=$(cat "${BUILDDIR}/${path}.sha256.txt")
+                if ! echo "${sha256} ${BUILDDIR}/${path}" | sha256sum -c --quiet -; then
+                    echo "SHA256 checksum for ${path} copied from ${ROOTDIR}/dist/csm-${CSM_BASE_VERSION}/${path} is incorrect, looks like file was corrupted in transit."
+                    exit 1
+                fi
                 echo "ok"
             else
                 echo "Not found ${path} in CSM base, will need to download."
@@ -88,4 +100,3 @@ for arch in "${CN_ARCH[@]}"; do
         process_file "${url}" "images/compute/$(basename "${url}")" "yes"
     done
 done
-
